requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.38.0
accelerate>=0.26.0
peft>=0.7.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=4.24.0
optimum>=1.14.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment management
python-dotenv>=1.0.0

# Unsloth for GRPO
unsloth>=0.4.0
flash-attn>=2.4.0

# Dataset handling
datasets>=2.15.0
jsonlines>=4.0.0
pyarrow>=14.0.0

# Tokenization
tiktoken>=0.5.0

# Evaluation and metrics
scikit-learn>=1.3.0
evaluate>=0.4.0

# API access
openrouter>=0.3.0
requests>=2.31.0
aiohttp>=3.9.0

# Experiment tracking
wandb>=0.16.0

# Documentation
pyyaml>=6.0.0
jinja2>=3.1.0

# Development tools
black>=23.10.0
isort>=5.12.0
pytest>=7.4.0 